Provide "simple" registry type.
Provide the Simple registry type that is designed to operate without a true OKAPI, such as a static GitHub pages website.

The simple design only performs a fetch of the exact JSON.
The remote server is expected to provide the `-latest` version because this simple type does not perform any complex versioning logic.

Add `spring-boot-autoconfigure` because the `AnyNestedCondition` needs to be added to support condition on OR.
This OR conditions allows for selecting either the `OkapiCondition` or the `SimpleCondition` as an `HttpCondition`.
The `HttpClient` bean now only needs a single `@Conditional` to handle either the OKAPI or the Simple registry types.
I am open to better solutions.

Add `spring-boot-starter-test` to enable the use of `ReflectionTestUtils`.
The `ReflectionTestUtils` is used to mock the `httpClient` during testing.

## Additional Notes

The README table doesn't need the `|` at the end.
Simplify this to avoid having to add a lot of spaces by removing the ending `|`.
The `|` can be added if requested.