Provide "simple" registry type. (#34)
* Provide "simple" registry type.

Provide the Simple registry type that is designed to operate without a true OKAPI, such as a static GitHub pages website.

The simple design only performs a fetch of the exact JSON.
The remote server is expected to provide the `-latest` version because this simple type does not perform any complex versioning logic.

Add `spring-boot-autoconfigure` because the `AnyNestedCondition` needs to be added to support condition on OR.
This OR conditions allows for selecting either the `OkapiCondition` or the `SimpleCondition` as an `HttpCondition`.
The `HttpClient` bean now only needs a single `@Conditional` to handle either the OKAPI or the Simple registry types.
I am open to better solutions.

Add `spring-boot-starter-test` to enable the use of `ReflectionTestUtils`.
The `ReflectionTestUtils` is used to mock the `httpClient` during testing.

## Additional Notes

The README table doesn't need the `|` at the end.
Simplify this to avoid having to add a lot of spaces by removing the ending `|`.
The `|` can be added if requested.

* Add simple registry entries to StringModuleRegistryParserTest.

* Add unit test for expections, findModuleDescriptor_negative_throwsException().

* Add tests to handle ternary case with prepareHttpRequest().

The path needs to be tested with `/` and without `/`.
Add a parameter to `simpleRegistry()` and then use parameterized tests to handle this.

* Add unit tests for handling the retryLoad() 504 requests.

* Add Simple registry testing to StringModuleRegistryParserTest().

The previous tests  for `StringModuleRegistryParserTest()` that I added are incomplete.

* Reduce redundancy in SimpleModuleDescriptorLoaderTest.

Use `@MethodSource` instead of `@ValueSource` to allow for multiple parameters.
Follow the `@ParameterizedTest` name format used by other tests.
Move the common mock responses into a shared `mockResponse()` method.

* Make retry an integer to allow for testing the second retry loop condition.

* Try to reduce the supposed code duplication reported by SonarQube.

SonarQube does not appear to highlight what it is considering duplication so I am having to guess here.

There are also some "duplicates" that are necessary because the mock has to be called twice with the same exact values.
These must remain.

* Add SimpleModuleRegistryTest unit tests.

* Add simple tests to ModuleRegistryProviderTest.

Follow the existing tests similar to how Okapi is being done.
I decided to append the simple test arguments to the end of the `pluginConfigProvider()` to avoid getting lost and confused from jumping around the code a lot.

* Remove not needed JSONException usage.

* Use explicit exceptions in SimpleModuleDescriptorLoader.

* Use explicit exceptions in SimpleModuleDescriptorLoader, when catching.

* Revert "Use explicit exceptions in SimpleModuleDescriptorLoader, when catching."

This reverts commit bbd9b7cad7562216bd5081393383c33e5e3e68bc.

Following this recommendation from SonarQube results in even bigger complaints about:
```
Either re-interrupt this method or rethrow the "InterruptedException" that can be caught here.
```

* Revert "Use explicit exceptions in SimpleModuleDescriptorLoader."

This reverts commit 32392f13691c0a51477a64eb97a70a9c55fb7a22.

Following this recommendation from SonarQube results in even bigger complaints about:
```
Either re-interrupt this method or rethrow the "InterruptedException" that can be caught here.
```

* Attempt to simplify the SimpleModuleDescriptorLoader.

I wanted to be as consistent with the existing code patterns from this project.
However, the SonarQube is complaining about duplications.

Change the logic and flow of the code in an attempt to address the duplications.
If this shows improvements then I may continue this practice with `SimpleModuleRegistry.java`.

* Attempt to reduce duplications in SimpleModuleRegistry.

There is not much to reduce here, except for maybe the `isValid()` method.
Reduce the return statements.